+++ Project Euler/Files/Classes/Question Objects/Questions 1-100/Questions 26-50/Questions 41-45/Question44.m
@@ -128,7 +128,7 @@
             // numerator, where the denominator is 6),
             if((currentRoot % 6) == 0){
               // Set the difference to be the current pentagonal number P_k.
-              difference = currentPentagonalNumber;
+              difference = (uint)currentPentagonalNumber;
               
               // Break out of the loop.
               break;
@@ -264,7 +264,7 @@
             // numerator, where the denominator is 6),
             if((currentRoot % 6) == 0){
               // Set the difference to be the current pentagonal number P_k.
-              difference = currentPentagonalNumber;
+              difference = (uint)currentPentagonalNumber;
               
               // Break out of the loop.
               break;
