+++ Project Euler/Files/Classes/Question Objects/Questions 1-100/Questions 76-100/Questions 76-80/Question79.m
@@ -272,7 +272,7 @@
   
   // Variable to hold the current digit index of the characters needed to be
   // accepted.
-  int currentDigitIndex = aCharacters.length - 1;
+  int currentDigitIndex = (int)aCharacters.length - 1;
   
   // Variable to hold the digit in the passcode we are looking at.
   uint digit = 0;
